All tests passed. Fixed bug in getting accounts

+++ TDAmeritrade.Client/Models/AccountsAndTrading/OrderActivity.cs
@@ -15,17 +15,17 @@
     {
         public OrderActivityType activityType { get; set; }
         public OrderActivityExecutionType executionType { get; set; }
-        public int quantity { get; set; }
-        public int orderRemainingQuantity { get; set; }
+        public float quantity { get; set; }
+        public float orderRemainingQuantity { get; set; }
         public Executionleg[] executionLegs { get; set; }
     }
 
     public class Executionleg
     {
-        public int legId { get; set; }
-        public int quantity { get; set; }
-        public int mismarkedQuantity { get; set; }
-        public int price { get; set; }
+        public long legId { get; set; }
+        public float quantity { get; set; }
+        public float mismarkedQuantity { get; set; }
+        public float price { get; set; }
         public string time { get; set; }
     }
